+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADPIMGroupSetting/MSFT_AADPIMGroupSetting.psm1
@@ -1544,13 +1544,13 @@ function Export-TargetResource
             }
         }
 
-        $batchResponses = Invoke-M365DSCGraphBatchRequest -Requests $batchRequests
+        $batchResponses = Invoke-M365DSCGraphBatchRequest -Requests $batchRequests -AsList
 
         $dscContent = ''
-        foreach ($response in $batchResponses)
+        foreach ($group in $Script:exportedGroups)
         {
+            $response = $batchResponses.Where({ $_.id -eq $group.Id })
             $getValue = $response.body.value
-            $group = $Script:exportedGroups | Where-Object -FilterScript { $_.Id -eq $response.id }
             Write-M365DSCHost -Message "    |---[$j/$($Script:exportedGroups.Count)] $($group.DisplayName)" -DeferWrite
 
             $i = 1