+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOMailboxCalendarConfiguration/MSFT_EXOMailboxCalendarConfiguration.psm1
@@ -280,7 +280,7 @@ function Get-TargetResource
             UseBrightCalendarColorThemeInOwa         = $config.UseBrightCalendarColorThemeInOwa
             WeatherEnabled                           = $config.WeatherEnabled
             WeatherLocationBookmark                  = $config.WeatherLocationBookmark
-            WeatherLocations                         = [Array]$config.WeatherLocations
+            WeatherLocations                         = [System.String[]]$config.WeatherLocations
             WeatherUnit                              = $config.WeatherUnit
             WeekStartDay                             = $config.WeekStartDay
             WorkDays                                 = $config.WorkDays